requirements.txt
gradio
langchain-openai
langchain_experimental
python-dotenv
langchain
langchain-huggingface
huggingface_hub
langchain-community
duckduckgo-search
